--- pkg/communication/README.md
@@ -1,6 +1,0 @@
-# Communication
-
-Handle things related to communication with market portal
-1. get jobs to run
-2. upload status
-3. ......
