--- pkg/job/README.md
@@ -1,4 +1,0 @@
-# job
-handle things related to job
-job is created by user , then scheduled by marketmanger , then cpodmanger get the job and run it.
-
