--- pkg/job/general-job/README.md
@@ -1,2 +1,0 @@
-# general job
-handle things related to general k8s job
