--- pkg/job/kubeflow-mpijob/README.md
@@ -1,2 +1,0 @@
-# kubeflow mpijob
-handle things related to kubeflow mpijob
