--- pkg/job/kubeflow-pytorchjob/README.md
@@ -1,2 +1,0 @@
-# kubeflow pytorchjob
-handle things related to kubeflow pytorchjob
