--- pkg/job/utils/README.md
@@ -1,1 +1,0 @@
-# utils for job
