+++ pkg/model-uploader/model_uploader.go
@@ -4,61 +4,13 @@ import (
 	"bufio"
 	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"os"
 	"sxwl/3k/pkg/config"
-	"sxwl/3k/pkg/job"
-	"sxwl/3k/pkg/job/state"
 	"sxwl/3k/pkg/log"
 	"sxwl/3k/pkg/storage"
-	"time"
-
-	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 )
-
-// moniter the given mpijob , until it stops (fail , finish , mannual interrupt or whatever)
-// 返回 (KUBEFLOW)MPIJob结束时的状态 status  和  监控过程中发生的程序必须中断不可恢复的错误 err
-// 当status为“”时，err必不为nil , status不为“”时，err必为nil。
-// 当status为“Complete”时，代表需要进一步执行模型上传工作。
-// 当err != nil 时，代表监控无法继续进行，需要由K8S触发重启。
-func UntilMPIJobFinish(mpiJobName string) (string, error) {
-	errCnt := 0
-	for {
-		s, err := job.GetState(config.CPOD_NAMESPACE, mpiJobName)
-		if err != nil {
-			//如果MPIJob已经被删除
-			if k8serrors.IsNotFound(err) {
-				log.SLogger.Infow("job is deleted")
-				//if job is deleted , do upload as well
-				return config.STATUS_NEEDS_UPLOAD_MODEL, nil
-				//return "deleted", nil
-			}
-			log.SLogger.Errorw("error when wait job finish", "error", err, "errcnt", errCnt)
-			errCnt += 1
-			//持续报错， 返回异常
-			if errCnt > 10 {
-				return "", errors.New("cant get state in 10 minutes")
-			}
-		} else { //errCnt置0
-			errCnt = 0
-		}
-		log.SLogger.Infow("job status", "jobname", mpiJobName, "status", s.JobStatus)
-		//判断MPIJob是否已经终止（状态不会再变）
-		if s.JobStatus.NoMoreChange() {
-			if s.JobStatus == state.JobStatusSucceed {
-				log.SLogger.Infow("job succeed")
-				return config.STATUS_NEEDS_UPLOAD_MODEL, nil
-			}
-			log.SLogger.Infow("job failed , nothing to upload")
-			return "nouploadneeded", nil
-		}
-		//继续等待
-		log.SLogger.Infow("keep waiting")
-		time.Sleep(time.Minute)
-	}
-}
 
 // err != nil 代表上传任务失败，程序无法继续执行。
 // 需要由K8S触发重启
