--- pkg/resource/README.md
@@ -1,5 +1,0 @@
-# resource
-collect cpod resource info 
-info mainly from two source 
-1. k8s( with client-go) ,  kubectl get nodes -ojson 
-2. node&dcgm exporter  or prometheus
